fix(AIMessage): harden audio block + add Storybook example
- Guard audioUrl against javascript: scheme (mirrors image/file blocks).
- Pass content.duration through as AudioPlayer fallbackDuration so the
  time shows before audio metadata loads. Addresses review on #285.
- Add WithAudioBlock story to AIMessage stories and document the audio
  block in the docs table, so the inline player is visible in Storybook.
- Extract the sample-audio WAV generator into a shared sampleAudio.ts
  reused by AudioPlayer and AIMessage stories (DRY).
