+++ src/components/AudioPlayer/AudioPlayer.stories.tsx
@@ -1,137 +1,11 @@
 import * as React from 'react';
 import type { Meta, StoryObj } from '@storybook/react-vite';
 import { AudioPlayer } from './AudioPlayer';
+import { getLongAudio, getSampleAudio, getShortAudio } from './sampleAudio';
 import { Card, CardContent, CardHeader, CardTitle } from '../Card';
 import { Text } from '../Text';
 import { Avatar } from '../Avatar';
 import { Badge } from '../Badge';
-
-// ============================================================================
-// Audio Sample Generation
-// ============================================================================
-
-/**
- * Creates a synthetic audio blob URL using Web Audio API.
- * This generates a short tone that works locally without CORS issues.
- */
-function createSampleAudioUrl(durationSec = 5, frequency = 440): string {
-  // Create audio context
-  const audioContext = new (
-    window.AudioContext ||
-    (window as typeof window & { webkitAudioContext?: typeof AudioContext })
-      .webkitAudioContext
-  )();
-  const sampleRate = audioContext.sampleRate;
-  const numSamples = Math.floor(sampleRate * durationSec);
-
-  // Create stereo buffer
-  const buffer = audioContext.createBuffer(2, numSamples, sampleRate);
-
-  // Generate a pleasant tone with envelope
-  for (let channel = 0; channel < 2; channel++) {
-    const data = buffer.getChannelData(channel);
-    for (let i = 0; i < numSamples; i++) {
-      const t = i / sampleRate;
-      // Apply envelope (fade in/out)
-      const envelope =
-        Math.min(t * 4, 1) * Math.min((durationSec - t) * 4, 1) * 0.3;
-      // Generate tone with harmonics
-      const fundamental = Math.sin(2 * Math.PI * frequency * t);
-      const harmonic1 = Math.sin(2 * Math.PI * frequency * 2 * t) * 0.5;
-      const harmonic2 = Math.sin(2 * Math.PI * frequency * 3 * t) * 0.25;
-      data[i] = envelope * (fundamental + harmonic1 + harmonic2);
-    }
-  }
-
-  // Convert to WAV format
-  const wavBuffer = audioBufferToWav(buffer);
-  const blob = new Blob([wavBuffer], { type: 'audio/wav' });
-  return URL.createObjectURL(blob);
-}
-
-/**
- * Converts an AudioBuffer to WAV format
- */
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
-function audioBufferToWav(buffer: any): ArrayBuffer {
-  const numChannels = buffer.numberOfChannels;
-  const sampleRate = buffer.sampleRate;
-  const format = 1; // PCM
-  const bitDepth = 16;
-
-  const bytesPerSample = bitDepth / 8;
-  const blockAlign = numChannels * bytesPerSample;
-
-  const samples = buffer.length;
-  const dataSize = samples * blockAlign;
-  const bufferSize = 44 + dataSize;
-
-  const arrayBuffer = new ArrayBuffer(bufferSize);
-  const view = new DataView(arrayBuffer);
-
-  // WAV header
-  const writeString = (offset: number, str: string) => {
-    for (let i = 0; i < str.length; i++) {
-      view.setUint8(offset + i, str.charCodeAt(i));
-    }
-  };
-
-  writeString(0, 'RIFF');
-  view.setUint32(4, bufferSize - 8, true);
-  writeString(8, 'WAVE');
-  writeString(12, 'fmt ');
-  view.setUint32(16, 16, true); // Subchunk1Size
-  view.setUint16(20, format, true);
-  view.setUint16(22, numChannels, true);
-  view.setUint32(24, sampleRate, true);
-  view.setUint32(28, sampleRate * blockAlign, true);
-  view.setUint16(32, blockAlign, true);
-  view.setUint16(34, bitDepth, true);
-  writeString(36, 'data');
-  view.setUint32(40, dataSize, true);
-
-  // Write interleaved audio data
-  let offset = 44;
-  for (let i = 0; i < samples; i++) {
-    for (let ch = 0; ch < numChannels; ch++) {
-      const sample = Math.max(-1, Math.min(1, buffer.getChannelData(ch)[i]));
-      view.setInt16(
-        offset,
-        sample < 0 ? sample * 0x8000 : sample * 0x7fff,
-        true
-      );
-      offset += 2;
-    }
-  }
-
-  return arrayBuffer;
-}
-
-// Create sample audio URLs (memoized to avoid regenerating)
-let _sampleAudioUrl: string | null = null;
-let _shortAudioUrl: string | null = null;
-let _longAudioUrl: string | null = null;
-
-function getSampleAudio(): string {
-  if (!_sampleAudioUrl) {
-    _sampleAudioUrl = createSampleAudioUrl(10, 440); // 10 seconds, A4 note
-  }
-  return _sampleAudioUrl;
-}
-
-function getShortAudio(): string {
-  if (!_shortAudioUrl) {
-    _shortAudioUrl = createSampleAudioUrl(3, 523.25); // 3 seconds, C5 note
-  }
-  return _shortAudioUrl;
-}
-
-function getLongAudio(): string {
-  if (!_longAudioUrl) {
-    _longAudioUrl = createSampleAudioUrl(30, 329.63); // 30 seconds, E4 note
-  }
-  return _longAudioUrl;
-}
 
 // ============================================================================
 // Meta
